functions/requirements.txt
fastai>=1.0

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
torchaudio==0.8.1

google-cloud-storage==1.38.0